requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.7.5
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.89.1
fastapi-socketio==0.0.9
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
python-jose
multidict==6.0.4
passlib==1.7.4
psycopg2-binary==2.9.5
pydantic==1.10.4
python-dotenv==0.21.1
python-engineio==4.3.4
python-socketio==5.7.2
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.0
starlette==0.22.0
typing-extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
python-multipart==0.0.5